Move utils to root folder
