+++ src/aiohubspace/util.py
